A bit reorganized drivers and fixed a couple of errors btw
