+++ lib/drivers/vmx/config.go
@@ -33,12 +33,12 @@ type Config struct {
 
 	// Alter allows you to control how much resources will be used:
 	// * Negative (<0) value will alter the total resource count before provisioning so you will be
-	//   able to save some resources for the host system (recommended -2 for CPU and -10240 for RAM
+	//   able to save some resources for the host system (recommended -2 for CPU and -10 for RAM
 	//   for disk caching)
 	// * Positive (>0) value could also be available (but check it in your vmware dist in advance)
 	//   Please be careful here - noone wants the VM to fail allocation because of that...
 	CpuAlter int `json:"cpu_alter"` // 0 do nothing, <0 reduces number available CPUs, >0 increases it (dangerous)
-	RamAlter int `json:"ram_alter"` // 0 do nothing, <0 reduces amount of available RAM (MB), >0 increases it (dangerous)
+	RamAlter int `json:"ram_alter"` // 0 do nothing, <0 reduces amount of available RAM (GB), >0 increases it (dangerous)
 
 	// Overbook options allows tenants to reuse the resources
 	// It will be used only when overbook is allowed by the tenants. It works by just adding those
@@ -47,7 +47,7 @@ type Config struct {
 	// to have virtually 28 CPUs. 3rd will not be running because 2 tenants will eat all 28 virtual
 	// CPUs. Same applies to the RamOverbook.
 	CpuOverbook uint `json:"cpu_overbook"` // How much CPUs could be reused by multiple tenants
-	RamOverbook uint `json:"ram_overbook"` // How much RAM (MB) could be reused by multiple tenants
+	RamOverbook uint `json:"ram_overbook"` // How much RAM (GB) could be reused by multiple tenants
 
 	DownloadUser     string `json:"download_user"`     // The user will be used in download operations
 	DownloadPassword string `json:"download_password"` // The password will be used in download operations
@@ -123,7 +123,7 @@ func (c *Config) Validate() (err error) {
 	if err != nil {
 		return err
 	}
-	ram_stat := mem_stat.Total / 1048576 // Getting MB from Bytes
+	ram_stat := mem_stat.Total / 1073741824 // Getting GB from Bytes
 
 	if c.RamAlter < 0 && int(ram_stat) <= -c.RamAlter {
 		return fmt.Errorf("VMX: |RamAlter| can't be more or equal the avaialble Host RAM: |%d| > %d", c.RamAlter, ram_stat)
