+++ lib/drivers/vmx/util.go
@@ -12,8 +12,6 @@
 
 package vmx
 
-// VMWare VMX (Fusion/Workstation) driver to manage VMs & images
-
 import (
 	"bufio"
 	"bytes"
@@ -29,220 +27,26 @@ import (
 	"time"
 
 	"github.com/hpcloud/tail"
-	"github.com/shirou/gopsutil/cpu"
-	"github.com/shirou/gopsutil/mem"
 
-	"github.com/adobe/aquarium-fish/lib/crypt"
 	"github.com/adobe/aquarium-fish/lib/drivers"
 	"github.com/adobe/aquarium-fish/lib/util"
 )
 
-// Implements drivers.ResourceDriver interface
-type Driver struct {
-	cfg Config
-
-	node_cpu uint // In logical threads
-	node_ram uint // In RAM megabytes
-}
-
-func init() {
-	drivers.DriversList = append(drivers.DriversList, &Driver{})
-}
-
-func (d *Driver) Name() string {
-	return "vmx"
-}
-
-func (d *Driver) IsRemote() bool {
-	return false
-}
-
-func (d *Driver) Prepare(config []byte) error {
-	if err := d.cfg.Apply(config); err != nil {
-		return err
-	}
-	if err := d.cfg.Validate(); err != nil {
-		return err
-	}
-
-	// Collect node resources status
-	cpu_stat, err := cpu.Counts(true)
-	if err != nil {
-		return err
-	}
-	d.node_cpu = uint(cpu_stat)
-
-	mem_stat, err := mem.VirtualMemory()
-	if err != nil {
-		return err
-	}
-	d.node_ram = uint(mem_stat.Total / 1048576) // Getting MB from Bytes
-
-	// TODO: Cleanup the image directory in case the images are not good
-
-	return nil
-}
-
-func (d *Driver) ValidateDefinition(definition string) error {
-	var def Definition
-	return def.Apply(definition)
-}
-
-func (d *Driver) DefinitionResources(definition string) drivers.Resources {
-	var def Definition
-	def.Apply(definition)
-
-	return def.Resources
-}
-
 // Returns the total resources available for the node after alteration
-func (d *Driver) getTotalResources() (total_cpu, total_ram uint) {
+func (d *Driver) getAvailResources() (avail_cpu, avail_ram uint) {
 	if d.cfg.CpuAlter < 0 {
-		total_cpu = d.node_cpu - uint(-d.cfg.CpuAlter)
+		avail_cpu = d.total_cpu - uint(-d.cfg.CpuAlter)
 	} else {
-		total_cpu = d.node_cpu + uint(d.cfg.CpuAlter)
+		avail_cpu = d.total_cpu + uint(d.cfg.CpuAlter)
 	}
 
 	if d.cfg.RamAlter < 0 {
-		total_ram = d.node_ram - uint(-d.cfg.RamAlter)
+		avail_ram = d.total_ram - uint(-d.cfg.RamAlter)
 	} else {
-		total_ram = d.node_ram + uint(d.cfg.RamAlter)
+		avail_ram = d.total_ram + uint(d.cfg.RamAlter)
 	}
 
 	return
-}
-
-// Allow Fish to ask the driver about it's capacity (free slots) of a specific definition
-func (d *Driver) AvailableCapacity(node_usage drivers.Resources, definition string) int64 {
-	var out_count int64
-
-	var req Definition
-	if err := req.Apply(definition); err != nil {
-		log.Println("VMX: Unable to apply definition:", err)
-		return -1
-	}
-
-	total_cpu, total_ram := d.getTotalResources()
-
-	// Check if the node has the required resources - otherwise we can't run it anyhow
-	if req.Resources.Cpu > total_cpu {
-		return 0
-	}
-	if req.Resources.Ram > total_ram {
-		return 0
-	}
-	// TODO: Check disk requirements
-
-	// Since we have the required resources - let's check if tenancy allows us to expand them to
-	// run more tenants here
-	if node_usage.IsEmpty() {
-		// In case we dealing with the first one - we need to set usage modificators, otherwise
-		// those values will mess up the next calculations
-		node_usage.Multitenancy = req.Resources.Multitenancy
-		node_usage.CpuOverbook = req.Resources.CpuOverbook
-		node_usage.RamOverbook = req.Resources.RamOverbook
-	}
-	if node_usage.Multitenancy && req.Resources.Multitenancy {
-		// Ok we can run more tenants, let's calculate how much
-		if node_usage.CpuOverbook && req.Resources.CpuOverbook {
-			total_cpu += d.cfg.CpuOverbook
-		}
-		if node_usage.RamOverbook && req.Resources.RamOverbook {
-			total_ram += d.cfg.RamOverbook
-		}
-	}
-
-	// Calculate how much of those definitions we could run
-	out_count = int64((total_cpu - node_usage.Cpu) / req.Resources.Cpu)
-	ram_count := int64((total_ram - node_usage.Ram) / req.Resources.Ram)
-	if out_count > ram_count {
-		out_count = ram_count
-	}
-	// TODO: Add disks into equation
-
-	return out_count
-}
-
-/**
- * Allocate VM with provided images
- *
- * It automatically download the required images, unpack them and runs the VM.
- * Not using metadata because there is no good interfaces to pass it to VM.
- */
-func (d *Driver) Allocate(definition string, metadata map[string]interface{}) (string, string, error) {
-	var def Definition
-	if err := def.Apply(definition); err != nil {
-		log.Println("VMX: Unable to apply definition:", err)
-		return "", "", err
-	}
-
-	// Generate unique id from the hw address and required directories
-	buf := crypt.RandBytes(6)
-	buf[0] = (buf[0] | 2) & 0xfe // Set local bit, ensure unicast address
-	vm_id := fmt.Sprintf("%02x%02x%02x%02x%02x%02x", buf[0], buf[1], buf[2], buf[3], buf[4], buf[5])
-	vm_hwaddr := fmt.Sprintf("%02x:%02x:%02x:%02x:%02x:%02x", buf[0], buf[1], buf[2], buf[3], buf[4], buf[5])
-
-	vm_network := def.Resources.Network
-	if vm_network == "" {
-		vm_network = "hostonly"
-	}
-
-	vm_dir := filepath.Join(d.cfg.WorkspacePath, vm_id)
-	vm_images_dir := filepath.Join(vm_dir, "images")
-
-	// Load the required images
-	img_path, err := d.loadImages(&def, vm_images_dir)
-	if err != nil {
-		return vm_hwaddr, "", err
-	}
-
-	// Clone VM from the image
-	vmx_path := filepath.Join(vm_dir, vm_id+".vmx")
-	args := []string{"-T", "fusion", "clone",
-		img_path, vmx_path,
-		"linked", "-snapshot", "original",
-		"-cloneName", vm_id,
-	}
-	if _, _, err := runAndLog(120*time.Second, d.cfg.VmrunPath, args...); err != nil {
-		return vm_hwaddr, "", err
-	}
-
-	// Change cloned vm configuration
-	if err := util.FileReplaceToken(vmx_path,
-		true, true, true,
-		"ethernet0.addressType =", `ethernet0.addressType = "static"`,
-		"ethernet0.address =", fmt.Sprintf("ethernet0.address = %q", vm_hwaddr),
-		"ethernet0.connectiontype =", fmt.Sprintf("ethernet0.connectiontype = %q", vm_network),
-		"numvcpus =", fmt.Sprintf(`numvcpus = "%d"`, def.Resources.Cpu),
-		"cpuid.corespersocket =", fmt.Sprintf(`cpuid.corespersocket = "%d"`, def.Resources.Cpu),
-		"memsize =", fmt.Sprintf(`memsize = "%d"`, def.Resources.Ram*1024),
-	); err != nil {
-		log.Println("VMX: Unable to change cloned VM configuration", vmx_path)
-		return vm_hwaddr, "", err
-	}
-
-	// Create and connect disks to vmx
-	if err := d.disksCreate(vmx_path, def.Resources.Disks); err != nil {
-		log.Println("VMX: Unable create disks for VM", vmx_path)
-		return vm_hwaddr, "", err
-	}
-
-	// Run the background monitoring of the vmware log
-	if d.cfg.LogMonitor {
-		go d.logMonitor(vmx_path)
-	}
-
-	// Run the VM
-	if _, _, err := runAndLog(120*time.Second, d.cfg.VmrunPath, "start", vmx_path, "nogui"); err != nil {
-		log.Println("VMX: Unable to run VM", vmx_path, err)
-		log.Println("VMX: Check the log info:", filepath.Join(filepath.Dir(vmx_path), "vmware.log"),
-			"and directory ~/Library/Logs/VMware/ for additional logs")
-		return vm_hwaddr, "", err
-	}
-
-	log.Println("VMX: Allocate of VM", vm_hwaddr, vmx_path, "completed")
-
-	return vm_hwaddr, "", nil
 }
 
 // Load images and returns the target image path for cloning
@@ -371,6 +175,7 @@ func (d *Driver) loadImages(def *Definition, vm_images_dir string) (string, erro
 	return target_path, nil
 }
 
+// Returns the path to VMX file from hwaddr identifier
 func (d *Driver) getAllocatedVMX(hwaddr string) string {
 	// Probably it's better to store the current list in the memory and
 	// update on fnotify or something like that...
@@ -410,6 +215,7 @@ func (d *Driver) getAllocatedVMX(hwaddr string) string {
 	return ""
 }
 
+// Creates VMDK disks described by the disks map
 func (d *Driver) disksCreate(vmx_path string, disks map[string]drivers.Disk) error {
 	// Create disk files
 	var disk_paths []string
@@ -570,6 +376,7 @@ func (d *Driver) disksCreate(vmx_path string, disks map[string]drivers.Disk) err
 	return nil
 }
 
+// Ensures the VM is not stale by monitoring the log
 func (d *Driver) logMonitor(vmx_path string) {
 	// Monitor the vmware.log file
 	log_path := filepath.Join(filepath.Dir(vmx_path), "vmware.log")
@@ -587,46 +394,7 @@ func (d *Driver) logMonitor(vmx_path string) {
 	log.Println("VMX: Done monitoring of log:", log_path)
 }
 
-func (d *Driver) Status(hwaddr string) string {
-	if len(d.getAllocatedVMX(hwaddr)) > 0 {
-		return drivers.StatusAllocated
-	}
-	return drivers.StatusNone
-}
-
-func (d *Driver) Snapshot(hwaddr string, full bool) (string, error) {
-	return "", fmt.Errorf("VMX: Snapshot not implemented")
-}
-
-func (d *Driver) Deallocate(hwaddr string) error {
-	vmx_path := d.getAllocatedVMX(hwaddr)
-	if len(vmx_path) == 0 {
-		log.Println("VMX: Unable to find VM with HW ADDR:", hwaddr)
-		return fmt.Errorf("VMX: No VM found with HW ADDR: %s", hwaddr)
-	}
-
-	// Sometimes it's stuck, so try to stop a bit more than usual
-	if _, _, err := runAndLogRetry(3, 60*time.Second, d.cfg.VmrunPath, "stop", vmx_path); err != nil {
-		log.Println("VMX: Unable to deallocate VM:", vmx_path)
-		return err
-	}
-
-	// Delete VM
-	if _, _, err := runAndLogRetry(3, 30*time.Second, d.cfg.VmrunPath, "deleteVM", vmx_path); err != nil {
-		log.Println("VMX: Unable to delete VM:", vmx_path)
-		return err
-	}
-
-	// Cleaning the VM images too
-	if err := os.RemoveAll(filepath.Dir(vmx_path)); err != nil {
-		return err
-	}
-
-	log.Println("VMX: Deallocate of VM", hwaddr, vmx_path, "completed")
-
-	return nil
-}
-
+// Runs & logs the executable command
 func runAndLog(timeout time.Duration, path string, arg ...string) (string, string, error) {
 	var stdout, stderr bytes.Buffer
